Fix mismatched parameter and return types

+++ VoyeurDataSource.m
@@ -86,7 +86,7 @@
     return document;
 }
 
-- (int)outlineView:(NSOutlineView *)outlineView numberOfChildrenOfItem:(id)item
+- (NSInteger)outlineView:(NSOutlineView *)outlineView numberOfChildrenOfItem:(id)item
 {
     VoyeurNode *node;
     NSArray *children;
@@ -104,7 +104,7 @@
     return fonts;
 }
 
-- (id)outlineView:(NSOutlineView *)outlineView child:(int)index ofItem:(id)item
+- (id)outlineView:(NSOutlineView *)outlineView child:(NSInteger)index ofItem:(id)item
 {
     VoyeurNode *node;
     NSArray *children;
